refactor : InputView, ResultView 리팩토링 InputView   - 메소드명 변경 및 scanner close 추가 ResultView   - Car list를 받아 프린트하도록 변경
