+++ src/main/java/step3/util/ResultView.java
@@ -1,20 +1,21 @@
-package step3;
+package step3.util;
+
+import step3.domain.Car;
+
+import java.util.List;
 
 public class ResultView {
 
     private final String GAME_RESULT = "실행결과";
     private final String DASH = "-";
 
-    public void printGameResult(int result) {
-
-        for(int i = 0; i < result; i++) {
-            System.out.print(DASH);
+    public void printGameResult(List<Car> cars) {
+        for(Car car : cars) {
+            System.out.println(car.getDistance());
         }
-
-        System.out.println();
     }
 
-    public void printGameResultText() {
+    public void printGameResultHeader() {
         System.out.println(GAME_RESULT);
     }
